+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/MvcConfig.java
@@ -1,9 +1,4 @@
-package camp.nextstep;
-
-import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapterRegistry;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecutionAdapter;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerMappingRegistry;
+package com.interface21.webmvc.servlet.mvc.tobe;
 
 public class MvcConfig {
